feat(qasm-playground): add support for default QASM code
This commit adds support for default QASM code in the QASMPlayground component. The component now accepts a `children` prop that can be used to pass in the initial QASM code. If no `children` prop is provided, a default QASM code will be used. The `QASMPlayground` component has been updated to set the initial value of `qasmCode` based on the `children` prop or the default QASM code.

fix(mdx-editor): Remove unused 'children' prop

This commit removes an unused 'children' prop from the MdxEditor component.